Made the PopUp Component component-agnostic for better future reuse
