+++ client/app/components/DraftPage/TheStash/StashItem.css
@@ -24,7 +24,6 @@
   position: absolute;
   top: -5px;
   cursor: pointer;
-  /* border: 1px solid #fff; */
   border: none;
   border-radius: 50%;
   color: #fff;
