+++ client/app/components/PopUp/PopUp.css
@@ -1,0 +1,65 @@
+#popup {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: calc(100% - 15rem);
+  height: calc(10% - 2rem);
+  box-shadow: 0 -1px 2px rgba(140, 156, 200, 0.5);
+  background-color: hsl(210, 54%, 22%);
+  padding: 1rem 5rem 1rem 10rem;
+  font-size: 1.125rem;
+}
+
+#close-popup {
+  position: absolute;
+  top: 5px;
+  right: 10px;
+  font-size: 1.25rem;
+  border: none !important;
+  background-color: transparent !important;
+  transition: text-shadow .3s, color .3s;
+  -webkit-transition: text-shadow .3s, color .3s;
+}
+
+#close-popup:hover {
+  color: hsl(20, 100%, 83%);
+  text-shadow: -1px 1px 0 rgba(255, 113, 40, .5),
+               -2px 2px 2px #000;
+  transition: text-shadow .3s, color .3s;
+  -webkit-transition: text-shadow .3s, color .3s;
+}
+
+.pop-up-line {
+  margin: 0;
+}
+
+@media (min-width: 1750px) {
+  #popup {
+    font-size: 1.75rem;
+  }
+}
+
+@media (max-width: 1000px) {
+  #popup {
+    font-size: 1rem;
+    width: calc(100% - 10rem);
+    height: calc(15% - 1rem);
+    padding: 0 5rem;
+  }
+}
+
+@media (max-width: 580px) {
+  #popup {
+    width: calc(100% - 4rem);
+    padding: .5rem 2rem;
+  }
+}
+
+@media (max-width: 360px) {
+  #popup {
+    height: calc(20% - 1rem);
+  }
+}
