+++ client/app/components/PopUp/PopUp.jsx
@@ -1,0 +1,36 @@
+import React, { useState, useEffect } from 'react';
+import './PopUp.css';
+
+const PopUp = ({ message }) => {
+  const [showPopUp, togglePopUp] = useState(true);
+
+  useEffect(() => {
+    togglePopUp(true);
+  }, [message])
+
+  let splitUpMessage = message.split('\n');
+
+  const handleClosePopUp = (e) => {
+    e.preventDefault();
+    togglePopUp(false);
+  }
+
+  if (!showPopUp) {
+    return (<></>);
+  }
+
+  return (
+    <div id='popup' >
+      <button
+        id='close-popup'
+        className='button-style'
+        aria-label='close popup button'
+        onClick={handleClosePopUp} >
+        <i className="fas fa-times"></i>
+      </button>
+      {splitUpMessage.map(line => <p className='pop-up-line' >{line}</p>)}
+    </div>
+  );
+};
+
+export default PopUp;
