+++ client/app/components/SignInPage/SignInPage.css
@@ -63,46 +63,6 @@
   transition: box-shadow .3s;
 }
 
-#temp-popup {
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  display: flex;
-  align-items: center;
-  width: calc(100% - 15rem);
-  height: calc(10% - 2rem);
-  box-shadow: 0 -1px 2px rgba(140, 156, 200, 0.5);
-  background-color: hsl(210, 54%, 22%);
-  padding: 1rem 5rem 1rem 10rem;
-  font-size: 1.125rem;
-}
-
-#signin-page-close-popup {
-  position: absolute;
-  top: 5px;
-  right: 10px;
-  font-size: 1.25rem;
-  border: none !important;
-  background-color: transparent !important;
-  transition: text-shadow .3s, color .3s;
-  -webkit-transition: text-shadow .3s, color .3s;
-}
-
-#signin-page-close-popup:hover {
-  color: hsl(20, 100%, 83%);
-  text-shadow: -1px 1px 0 rgba(255, 113, 40, .5),
-               -2px 2px 2px #000;
-  transition: text-shadow .3s, color .3s;
-  -webkit-transition: text-shadow .3s, color .3s;
-}
-
-
-@media (min-width: 1750px) {
-  #temp-popup {
-    font-size: 1.75rem;
-  }
-}
-
 @media (max-width: 1000px) {
   #signin-page-container {
     grid-template-columns: 100%;
@@ -121,24 +81,4 @@
     margin: 0;
   }
 
-  #temp-popup {
-    font-size: 1rem;
-    width: calc(100% - 10rem);
-    height: calc(15% - 1rem);
-    padding: 0 5rem;
-  }
-
-}
-
-@media (max-width: 580px) {
-  #temp-popup {
-    width: calc(100% - 4rem);
-    padding: .5rem 2rem;
-  }
-}
-
-@media (max-width: 360px) {
-  #temp-popup {
-    height: calc(20% - 1rem);
-  }
 }
