+++ client/app/components/SignInPage/SignInPage.jsx
@@ -2,20 +2,12 @@ import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import SignInForm from './SignInForm/SignInForm.jsx';
 import SignUpForm from './SignUpForm/SignUpForm.jsx';
-import PopUp from './PopUp.jsx';
+import PopUp from '../PopUp/PopUp.jsx';
 import './SignInPage.css';
 
 const SignInPage = ({ inOrUp }) => {
-  const [showPopUp, togglePopUp] = useState(true);
 
-  useEffect(() => {
-    togglePopUp(true);
-  }, [inOrUp])
-
-  const handleClosePopUp = (e) => {
-    e.preventDefault();
-    togglePopUp(false);
-  }
+  const popUpMessage = `This page is currently only for show and has no backend infrastructure to support it.\nYou may try to sign ${inOrUp === 'in' ? 'in' : 'up'} if you'd like, but nothing would happen if this pop up appeared when you loaded the page.`
 
   return (
     <div id='signin-page-container' >
@@ -30,7 +22,7 @@ const SignInPage = ({ inOrUp }) => {
             : <SignUpForm />
         }
       </div>
-      { showPopUp && <PopUp handleClosePopUp={handleClosePopUp} inOrUp={inOrUp} /> }
+      <PopUp message={popUpMessage} inOrUp={inOrUp} />
     </div>
   );
 };
